Refactor visible labels into labels hook
