+++ src/calendar/use-calendar-labels.ts
@@ -3,15 +3,18 @@
 
 import { CalendarProps } from './interfaces';
 import { useInternalI18n } from '../i18n/context.js';
+import { getDateLabel, renderMonthAndYear } from './utils/intl';
 
 export default function useCalendarLabels({
   granularity,
+  locale,
   i18nStrings,
   previousMonthAriaLabel,
   nextMonthAriaLabel,
   todayAriaLabel,
 }: {
   granularity: CalendarProps.Granularity;
+  locale: string;
   i18nStrings?: CalendarProps.I18nStrings;
   previousMonthAriaLabel?: string;
   nextMonthAriaLabel?: string;
@@ -33,5 +36,25 @@ export default function useCalendarLabels({
     ? i18n('currentMonthAriaLabel', i18nStrings?.currentMonthAriaLabel)
     : i18n('todayAriaLabel', i18nStrings?.todayAriaLabel ?? todayAriaLabel);
 
-  return { previousButtonLabel, nextButtonLabel, currentDateLabel };
+  const renderDate = (date: Date) =>
+    isMonthPicker ? date.toLocaleString(locale, { month: 'short' }) : date.getDate().toString();
+
+  const renderDateAnnouncement = (date: Date, isCurrentDate: boolean) => {
+    const formattedDate = isMonthPicker ? renderMonthAndYear(locale, date) : getDateLabel(locale, date, 'short');
+    if (isCurrentDate && currentDateLabel) {
+      return formattedDate + '. ' + currentDateLabel;
+    }
+    return formattedDate;
+  };
+
+  const renderHeaderText = (date: Date) =>
+    isMonthPicker ? date.getFullYear().toString() : renderMonthAndYear(locale, date);
+
+  return {
+    previousButtonLabel,
+    nextButtonLabel,
+    renderDate,
+    renderDateAnnouncement,
+    renderHeaderText,
+  };
 }
